Fixed issue with the filter as this was not taking into consideration URLs with a context path.
Solution consisted of stripping the context path from the new url before sending this to the request dispatcher.
Updated the pom.xml with the path to the sources directory as the project does not follow the standard maven structure and as a result the project was not being packaged correctly.